wip: pass refactor
- Move per-pass boilerplate to each pass's file
- delete KokkosPasses.cpp
- Make LoopMapping a single pass with option again
  - but with 2 different patterns (regular vs. team-level)
- Delete mem space assignment (for now)
