+++ mlir/include/lapis/Dialect/Kokkos/Transforms/Passes.h
@@ -20,20 +20,13 @@ namespace mlir {
 #define GEN_PASS_DECL
 #include "lapis/Dialect/Kokkos/Transforms/Passes.h.inc"
 
-void populateMemrefResultsToParamsPatterns(RewritePatternSet &patterns);
 std::unique_ptr<Pass> createMemrefResultsToParamsPass();
 
-void populateMemrefToKokkosScratchPatterns(RewritePatternSet &patterns);
 std::unique_ptr<Pass> createMemrefToKokkosScratchPass();
 
-void populateParallelUnitStepPatterns(RewritePatternSet &patterns);
 std::unique_ptr<Pass> createParallelUnitStepPass();
 
-void populateKokkosLoopMappingPatterns(RewritePatternSet &patterns, bool teamLevel);
 std::unique_ptr<Pass> createKokkosLoopMappingPass(bool teamLevel = false);
-
-void populateKokkosMemorySpaceAssignmentPatterns(RewritePatternSet &patterns);
-std::unique_ptr<Pass> createKokkosMemorySpaceAssignmentPass();
 
 std::unique_ptr<Pass> createKokkosDualViewManagementPass();
