+++ mlir/lib/Dialect/Kokkos/Transforms/CMakeLists.txt
@@ -1,11 +1,8 @@
 add_mlir_dialect_library(MLIRKokkosTransforms
-  KokkosPasses.cpp
-
   MemrefResultsToParams.cpp
   MemrefToKokkosScratch.cpp
   ParallelUnitStep.cpp
   KokkosLoopMapping.cpp
-  KokkosMemorySpaceAssignment.cpp
   KokkosDualViewManagement.cpp
 
   ADDITIONAL_HEADER_DIRS
