+++ mlir/lib/Dialect/Kokkos/Transforms/MemrefResultsToParams.cpp
@@ -9,6 +9,11 @@
 #include "mlir/IR/IRMapping.h"
 #include "mlir/IR/Matchers.h"
 
+namespace mlir {
+#define GEN_PASS_DEF_MEMREFRESULTSTOPARAMS
+#include "lapis/Dialect/Kokkos/Transforms/Passes.h.inc"
+}
+
 using namespace mlir;
 
 namespace {
@@ -28,5 +33,24 @@ struct MemrefResultsToParamsRewriter : public OpRewritePattern<func::FuncOp> {
 
 void mlir::populateMemrefResultsToParamsPatterns(RewritePatternSet &patterns) {
   patterns.add<MemrefResultsToParamsRewriter>(patterns.getContext());
+}
+
+struct MemrefResultsToParamsPass
+    : public impl::MemrefResultsToParamsBase<MemrefResultsToParamsPass> {
+
+  MemrefResultsToParamsPass() = default;
+  MemrefResultsToParamsPass(const MemrefResultsToParamsPass& pass) = default;
+
+  void runOnOperation() override {
+    auto *ctx = &getContext();
+    RewritePatternSet patterns(ctx);
+    populateMemrefResultsToParamsPatterns(patterns);
+    (void) applyPatternsAndFoldGreedily(getOperation(), std::move(patterns));
+  }
+};
+
+std::unique_ptr<Pass> mlir::createMemrefResultsToParamsPass()
+{
+  return std::make_unique<MemrefResultsToParamsPass>();
 }
